+++ src/app/pages/mesas/mesa-detalles/documentos-mesa/documentos-mesa.component.html
@@ -1,0 +1,46 @@
+<div class="w-full flex flex-col bg-white border border-solid border-grey-200">
+		<div class="flex items-center justify-between px-4 py-3 border border-solid border-grey-200">
+			<h5 class="m-0 text-base font-semibold">
+                @if(tipo === 0) { Sesiones }
+                    @if(tipo === 1) { AM }
+            </h5>
+			@if (mesa.estadoRegistroNombre != 'cerrado') {
+				<div class="rounded border border-sky-700 text-sky-700 px-3 py-2 cursor-pointer hover:bg-sky-700 hover:text-white" (click)="crearDocumento()">
+					Nueva
+                    @if(tipo === 0) { sesión }
+                    @if(tipo === 1) { am }
+				</div>
+			}
+		</div>
+		<div class="flex border-solid ">
+			<nz-table [nzData]="documentos()" class="w-full" [nzFrontPagination]="false" [nzShowPagination]="false" >
+				<thead>
+					<tr>
+						<th nzAlign="center" nzWidth="64px">Nro</th>
+						<th nzAlign="center">Nombre archivo</th>
+						<th nzAlign="center">Usuario</th>
+						<th nzAlign="center">Fecha</th>
+						<th nzAlign="center" nzWidth="80px"></th>
+					</tr>
+				</thead>
+				<tbody>
+					@for(detalle of documentos(); track $index){
+					<tr>
+						<td nzAlign="center">{{$index + 1}}</td>
+						<td nzAlign="center">{{detalle.nombre}}</td>
+						<td nzAlign="center">{{detalle.responsable}}</td>
+						<td nzAlign="center">{{detalle.fechaCreacion}}</td>
+						<td nzAlign="center">
+							<div class="flex justify-center gap-1">
+								<app-boton-descargar [rutaArchivo]="detalle.archivo"></app-boton-descargar>
+								<button nz-button nzDanger nzType="link" nz-tooltip nzTooltipTitle="Eliminar" (click)="deleteDocumento(detalle)">
+									<span nz-icon nzType="delete" nzTheme="outline"></span>
+								</button>
+							</div>
+						</td>
+					</tr>
+					}
+				</tbody>
+			</nz-table>
+		</div>
+	</div>
