+++ src/app/pages/mesas/mesas.component.ts
@@ -82,6 +82,7 @@ export default class MesasComponent {
 
   getParams() {
     this.route.queryParams.subscribe(params => {
+      this.loading = true
       if (Object.keys(params).length > 0) {        
         let campo = params['campo'] ?? 'mesaId'
 
@@ -119,6 +120,7 @@ export default class MesasComponent {
   obtenerMesasService(){
     this.mesasService.ListarMesas(this.pagination)
       .subscribe( resp => {
+        this.loading = false
         this.mesas.set(resp.data)
         this.pagination.total = resp.info?.total
       })
